Add key features
- [] Being able to use interactive mode
- [] Added several examples

+++ docker-sql.sh
@@ -1,44 +1,44 @@
 #!/bin/bash
 
-# Check for SQL file and image version arguments
-if [ "$#" -ne 2 ]; then
-    echo "Usage: $0 /path/to/query.sql sql-db-keyword"
-    exit 1
-fi
-
-source utils.sh
-# Define a constant for the container name
+# Define constants
 CONTAINER_NAME="sql_container"
 
-# Assign arguments to variables
-SQL_FILE=$1
-SECTION=$2
+# Source utility functions
+source utils.sh
 
-# Fetch configuration details from db.yaml for the given section
-SQL_CMD=$(get_config $SECTION "SQL_CMD" "db.yaml")
-DOCKER_ENV=$(get_config $SECTION "DOCKER_ENV" "db.yaml")
-VERSION=$(get_config $SECTION "VERSION" "db.yaml")
+# Fetch configuration details
+fetch_config $SECTION
 
-# Construct image name from section and version
-IMAGE_NAME=$SECTION:$VERSION
-
-# Check if essential variables are set, if not exit with an error
-if [ -z "$SQL_CMD" ] ; then
+# Check if SQL command is provided
+if [ -z "$SQL_CMD" ]; then
     echo "Please specify a SQL command to run on db.yaml"
     exit 1
 fi
 
 # Run SQL server in Docker container
-docker run --name $CONTAINER_NAME -e $DOCKER_ENV -d $IMAGE_NAME
+if [ "$MODE" == "client" ]; then
+    SQL_CMD=$(replace_path "$SQL_CMD" " ")
+    docker run --name $CONTAINER_NAME -e SQL_CMD="$SQL_CMD" -e $DOCKER_ENV -v $DIR:/data -d $IMAGE_NAME
+    
+    # Add alias to bashrc inside the container
+    alias="alias run-query='$SQL_CMD'"
+    docker exec $CONTAINER_NAME /bin/bash -c "echo \"$alias\" >> ~/.bashrc"
+    
+    sleep 10
+    docker exec -it $CONTAINER_NAME bash
+else
+    SQL_PATH="/query.sql"
+    SQL_CMD=$(replace_path "$SQL_CMD" "$SQL_PATH")
+    
+    docker run --name $CONTAINER_NAME -e $DOCKER_ENV -d $IMAGE_NAME
+    
+    # Wait for SQL server to initialize
+    sleep 10
 
-# Wait for SQL server to initialize
-sleep 10
+    # Copy the SQL file into the container and execute it
+    docker cp $SQL_FILE $CONTAINER_NAME:/query.sql
+    docker exec $CONTAINER_NAME sh -c "$SQL_CMD"
 
-# Copy the SQL file into the container
-docker cp $SQL_FILE $CONTAINER_NAME:/query.sql
-
-# Execute SQL script in SQL server inside the Docker container
-docker exec $CONTAINER_NAME sh -c "$SQL_CMD"
-
-# Cleanup: Stop and remove the container
-docker stop $CONTAINER_NAME && docker rm $CONTAINER_NAME
+    # Cleanup: Stop and remove the container
+    docker stop $CONTAINER_NAME && docker rm $CONTAINER_NAME
+fi
